backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
chromadb
GitPython
langchain
google-search-results
gunicorn
langchain-astradb
langchain-google-genai